fix(release): apply publish manifest transform before npm reads it
The heavy `optionalDependencies` (webdriverio, appium-*-driver) were being
moved into a custom `ddRuntimeDependencies` field by a `prepack` lifecycle
hook so a default `npm i` would not drag them in. But `npm publish` builds
the registry packument (which `npm install` resolves dependencies from) by
reading package.json *before* lifecycle scripts run, then packs the tarball
*after* them. So the prepack edit reached the tarball but never the
packument — installs still pulled the heavy tree and emitted its deprecated
transitive warnings (glob@10.5.0, whatwg-encoding@3.1.1).

Move the transform out of the prepack hook and apply it on disk in
publish-staged-release.js *before* `npm publish` is invoked, restoring the
original manifest afterward (the common-workspace publish needs `workspaces`,
and @semantic-release/git commits package.json after this script). Add a
guardrail that fails the release if the published version still lists
optionalDependencies.

- scripts/prepack.js -> scripts/publish-manifest.js: keep the pure, tested
  transformForPublish(); drop the lifecycle/backup machinery.
- scripts/postpack.js: deleted (no longer needed).
- package.json: remove the now-dead prepack/postpack scripts.
- tests/comments: rename prepack.test.js, refresh wording.
