+++ test/publish-manifest.test.js
@@ -1,4 +1,4 @@
-import { transformForPublish } from "../scripts/prepack.js";
+import { transformForPublish } from "../scripts/publish-manifest.js";
 import { createRequire } from "node:module";
 
 const require = createRequire(import.meta.url);
@@ -8,7 +8,7 @@ before(async function () {
   global.expect = expect;
 });
 
-describe("scripts/prepack transformForPublish", function () {
+describe("scripts/publish-manifest transformForPublish", function () {
   it("drops the workspaces field", function () {
     const out = transformForPublish({ workspaces: ["src/common"], name: "x" });
     expect(out).to.not.have.property("workspaces");
